tail refactor
